+++ CHANGELOG.md
@@ -6,8 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Consolidate Certbot Logs
+
+### Added
+- Consolidate Certbot Logs to own docker volume
+- Logger Module for dynamic logs
+### Changed
 - Switch the UI Logic from String-based comparison to id comparison
+### Fixed
+- Lost in Translation
+
+
 ## [0.4.9] - 2025-05-01
 ### Added
 - Develop Branch